+++ src/main/java/com/revolvingmadness/sculk/language/builtins/classes/types/ObjectType.java
@@ -1,5 +1,6 @@
 package com.revolvingmadness.sculk.language.builtins.classes.types;
 
+import com.revolvingmadness.sculk.language.ErrorHolder;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinClass;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinMethod;
 import com.revolvingmadness.sculk.language.builtins.classes.BuiltinType;
@@ -12,13 +13,26 @@ import java.util.List;
 public class ObjectType extends BuiltinType {
     public ObjectType() {
         super("Object", null);
+
         this.typeVariableScope.declare(List.of(TokenType.CONST), "init", new Init());
+        this.typeVariableScope.declare(List.of(TokenType.CONST), "toString", new ToString());
     }
 
     private static class Init extends BuiltinMethod {
         @Override
         public BuiltinClass call(Interpreter interpreter, List<BuiltinClass> arguments) {
             return new NullInstance();
+        }
+    }
+
+    private static class ToString extends BuiltinMethod {
+        @Override
+        public BuiltinClass call(Interpreter interpreter, List<BuiltinClass> arguments) {
+            if (arguments.size() != 0) {
+                throw ErrorHolder.invalidArgumentCount("toString", 0, arguments.size());
+            }
+
+            return this.boundClass.toStringMethod();
         }
     }
 }