+++ src/main/java/com/revolvingmadness/sculk/language/errors/NumberFormatError.java
@@ -1,0 +1,7 @@
+package com.revolvingmadness.sculk.language.errors;
+
+public class NumberFormatError extends Error {
+    public NumberFormatError(String message) {
+        super(message);
+    }
+}